Remove internal visitorId assertion
This permits migration from the upstream package.

We still check the user-supplied visitorId format but allow Matomo
server to deal with the incorrect format from the old package
(substring of sha1 digest).

The existing (UUID) user ID is also still supplied to Matomo because
we default to using the visitorId for that field.